make proper identifiers for loggers
Identifier for loggers (the part after logger. until a keyword like
name or level) must not contain periods.
Otherwise the property configuration parser fails because it already
interprets the part until the first period (e.g. "de") as a keyword.
It then fails with something like "logger de does not have a name".

Removed the output for logger additivity and appenders because
the defaults are normally enough.